+++ src/discord/index.js
@@ -1,0 +1,2 @@
+require("./client.js");
+require("./messages.js");
